+++ pillLaw_backend/src/main/java/com/pilllaw/pilllaw/entity/product/ProductPrice.java
@@ -25,7 +25,7 @@ public class ProductPrice extends BaseEntity {
     private Product product;
 
     @Column(nullable = false)
-    private Double price;
+    private Long price;
 
-    private Double salePrice;
+    private Long salePrice;
 }
